Refactored modules, added css variables
