+++ index.js
@@ -2,17 +2,17 @@ import Viewer from 'bpmn-js/lib/Viewer';
 
 import MoveCanvasModule from 'diagram-js/lib/navigation/movecanvas';
 import ZoomScrollModule from 'diagram-js/lib/navigation/zoomscroll';
-import callActivityDrilldownModule from './lib/customDrilldown/callActivityDrilldown';
-import drilldownCenteringModule from './lib/customDrilldown/drilldownCentering';
-import styleModule from './lib/styleModule';
+import callActivityModule from './modules/callActivityModule';
+import drilldownCentering from './modules/drilldownCentering';
+import styleModule from './modules/styleModule';
 
 var bpmnViewer = {
   Viewer: Viewer,
   customModules: {
     MoveCanvasModule,
     ZoomScrollModule,
-    callActivityDrilldownModule,
-    drilldownCenteringModule,
+    drilldownCentering,
+    callActivityModule,
     styleModule
   },
 };
