--- lib/customDrilldown/callActivityDrilldown/index.js
@@ -1,6 +1,0 @@
-import CallActivityDrilldown from './CallActivityDrilldown';
-
-export default {
-  __init__: ['callActivityDrilldown'],
-  callActivityDrilldown: ['type', CallActivityDrilldown],
-};
