+++ modules/callActivityModule/index.js
@@ -1,0 +1,6 @@
+import CallActivityModule from './CallActivityModule';
+
+export default {
+  __init__: ['callActivityModule'],
+  callActivityModule: ['type', CallActivityModule],
+};
