+++ modules/styleModule/StyleModule.js
